Move dryRunCheckbox out of aws-common-properties (#247)
