+++ packages/openops/src/index.ts
@@ -57,3 +57,4 @@ export * from './lib/azure/auth';
 export * from './lib/azure/subscription/get-subscription-dropdown';
 
 export * from './lib/axios-wrapper';
+export * from './lib/dry-run-property';
