+++ packages/openops/src/lib/dry-run-property.ts
@@ -1,0 +1,9 @@
+import { Property } from '@openops/blocks-framework';
+
+export function dryRunCheckBox(): any {
+  return Property.Checkbox({
+    displayName: 'Dry Run',
+    description: 'Run the operation in dry run mode',
+    required: false,
+  });
+}
